+++ PSFramework/internal/tepp/scripts/tepp-scriptblockname.ps1
@@ -1,0 +1,3 @@
+﻿Register-PSFTeppScriptblock -Name 'PSFramework-tepp-scriptblockname' -ScriptBlock {
+	[PSFramework.TabExpansion.TabExpansionHost]::Scripts.Keys
+}